Fixed Pathfinding issues
Created custom distance function to calculate distance in tiles not in world location.

+++ Source/TurnBasedStratagey/MyTile.h
@@ -27,7 +27,7 @@ public:
 		int J;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Pathfinding")
-		int F = 0;
+		float F = 0;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Pathfinding")
 		int G = 0;
